Fix LMDB register accesses
utils::regExists closes the LMDB when it exits. Whenever the db_res was
used, it was referring to a dangling pointer to a (non-existent)
memory-mapped address.

utils::regExists has been modified to return the content of the LMDB key
as a std::string if it exists, avoiding the previsouly described issue.

During the adaptation of utils::writeRawAddress, utils::readRawAddress,
utils utils::writeRawReg, utils::readRawReg, utils::writeReg and
utils::readReg, the following internal functions have been removed:

* utils::writeAddress
* utils::readAddress

Basic const-ification has been applied on the way.